End the start queue when startid is not used
Consider some singletons, or applications that start another program.
They generally don't show windows. There is no end to the start queue,
and the startid exposed by QPA in Qt indicates this type of application
processing. It is now handled uniformly in the platform plugin dxcb.

It is worth noting that this type of application needs to ensure that the
 Application object is released when it exits.

Log: Fixed some singletons not ending the startup queue