fix: deepin-movie's window is resized when moving mouse beside the edge
Drag a file into the window and drop, the mouse state should be from Leftbutton to NoButton.
But the current mouse state is always Leftbutton, which causes the window to be resized when moving mouse beside the edge.
Update the mouse buttons' state by querying the state from xserver.

Issue: https://github.com/linuxdeepin/dtk/issues/63